superadmin controller para proveedor

+++ TelExpress/src/main/java/com/example/telexpress/repository/ProveedorRepository.java
@@ -2,9 +2,10 @@ package com.example.telexpress.repository;
 
 
 import com.example.telexpress.entity.Proveedor;
+import jakarta.persistence.criteria.CriteriaBuilder;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ProveedorRepository extends JpaRepository<Proveedor,Long> {
+public interface ProveedorRepository extends JpaRepository<Proveedor, Integer> {
 }
